Separate classes into diferrent files.
